+++ src/mmop-worker.lisp
@@ -6,7 +6,8 @@
            start-node-v0
            start-node-v0-type
            start-node-v0-recipe
-           make-start-node-success-v0))
+           make-start-node-success-v0
+           make-start-node-failure-v0))
 (in-package :monomyth/mmop-worker)
 
 (defstruct (worker-ready-v0 (:constructor make-worker-ready-v0 ()))
@@ -25,6 +26,15 @@
   (type (error "type must be set") :read-only t))
 (defmethod create-frames ((message start-node-success-v0))
   `(,*mmop-v0* "START-NODE-SUCCESS" ,(start-node-success-v0-type message)))
+
+(defstruct (start-node-failure-v0
+            (:constructor make-start-node-failure-v0 (type reason)))
+  "MMOP/0 start-node-failure"
+  (type (error "type must be set") :read-only t)
+  (reason (error "reason must be set") :read-only t))
+(defmethod create-frames ((message start-node-failure-v0))
+  `(,*mmop-v0* "START-NODE-FAILURE" ,(start-node-failure-v0-type message)
+               ,(start-node-failure-v0-reason message)))
 
 (defun pull-worker-message (socket)
   "pulls down a message designed for the worker dealer socket and attempts to